% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-setup.R
\name{add_columns_if_missing}
\alias{add_columns_if_missing}
\title{Add Missing Columns}
\usage{
add_columns_if_missing(con)
}
\arguments{
\item{con}{Database connection}
}
\description{
Add Missing Columns
}
