% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-setup.R
\name{create_database_triggers}
\alias{create_database_triggers}
\title{Create Database Triggers}
\usage{
create_database_triggers(con)
}
\arguments{
\item{con}{Database connection}
}
\description{
Create Database Triggers
}
