% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-setup.R
\name{create_fts_table_if_missing}
\alias{create_fts_table_if_missing}
\title{Create FTS Table If Missing}
\usage{
create_fts_table_if_missing(con)
}
\arguments{
\item{con}{Database connection}
}
\description{
Create FTS Table If Missing
}
