% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-setup.R
\name{create_optimization_indexes}
\alias{create_optimization_indexes}
\title{Create Optimization Indexes}
\usage{
create_optimization_indexes(con)
}
\arguments{
\item{con}{Database connection}
}
\description{
Create Optimization Indexes
}
