% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-setup.R
\name{create_standard_schema}
\alias{create_standard_schema}
\title{Create Standard Schema}
\usage{
create_standard_schema(con)
}
\arguments{
\item{con}{Database connection}
}
\description{
Create Standard Schema
}
