% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debate_metadata_1850.R
\docType{data}
\name{debate_metadata_1850}
\alias{debate_metadata_1850}
\title{The 19th-century British Parliamentary debates metadata for the decade 1850}
\format{
A data frame with 985755 rows and 5 variables:

sentence_id
A unique ID assigned to each sentence of the corpus, taken from the digitized parliamentary debates.

speechdate
The date on which the sentence was spoken in year, month, day format.

debate
The title of the debate.

speaker
The recorded name of the person who spoke the sentence.

disambig_speaker
The disambiguated name of the person who spoke the sentence.
}
\source{
\href{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/ZCYJH8}{Harvard Dataverse}
}
\usage{
debate_metadata_1850
}
\description{
The Hansard corpus debate metadata retains information about the debates such as the debate name and the speaker.
The variables are as follows:
}
\examples{
data(debate_metadata_1850)

}
\references{
Buongiorno, Steph; Kalescky, Robert; Godat, Eric; Cerpa, Omar Alexander; Guldi, Jo (2021)
(\href{https://doi.org/10.7910/DVN/ZCYJH8})
}
\keyword{datasets}
