% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debate_metadata_1860.R
\docType{data}
\name{debate_metadata_1860}
\alias{debate_metadata_1860}
\title{The 19th-century British Parliamentary debates metadata for the decade 1860}
\format{
A data frame with 1035152 rows and 5 variables:

sentence_id
A unique ID assigned to each sentence of the corpus, taken from the digitized parliamentary debates.

speechdate
The date on which the sentence was spoken in year, month, day format.

debate
The title of the debate.
}
\source{
\\href{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/ZCYJH8}{Harvard Dataverse}
}
\usage{
debate_metadata_1860
}
\description{
The Hansard corpus debate metadata retains information about the debates such as the debate title and the speech date.
The variables are as follows:
}
\examples{
data(debate_metadata_1860)

}
\references{
Buongiorno, Steph; Kalescky, Robert; Godat, Eric; Cerpa, Omar Alexander; Guldi, Jo (2021)
(\\href{https://doi.org/10.7910/DVN/ZCYJH8})
}
\keyword{datasets}
