\name{download_hansard}
\alias{download_hansard}
\title{Download the full c19 Hansard corpus}
\usage{
  download_hansard()
}
\value{
  The full sized RData files of the c19 Hansard Corpus.
}
\description{
  Read cleaned, decade subsets of the 19th-century British Parliamentary debates, also known as Hansard.
}
\examples{
  download_hansard()
}