% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\docType{data}
\name{events}
\alias{events}
\title{The 19th-century British Parliamentary debates metadata for the decade 1840}
\format{
A data frame with 48485 rows and 4 variables:

event_name
The name of the event

scholar_assigned_date
The date assigned to the event by a historian.

period
The time period in which the event was mentioned.

occurances
The number of times the event was stated in a period.
}
\usage{
events
}
\description{
The Hansard corpus debate metadata retains information about the debates such as the debate name and the speaker.
The variables are as follows:
}
\examples{
data(events)

}
\references{
Buongiorno, Steph; Kalescky, Robert; Godat, Eric; Cerpa, Omar Alexander; Guldi, Jo (2021)
(\href{https://doi.org/10.7910/DVN/ZCYJH8})
}
\keyword{datasets}
