% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-search.R
\name{get_content_statistics}
\alias{get_content_statistics}
\title{Get Content Statistics}
\usage{
get_content_statistics(con)
}
\arguments{
\item{con}{Database connection}
}
\value{
List with content statistics
}
\description{
Get statistics about speech content and search capabilities
}
