% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-queries.R
\name{get_database_stats}
\alias{get_database_stats}
\title{Get Database Statistics}
\usage{
get_database_stats(con)
}
\arguments{
\item{con}{Database connection}
}
\value{
List with database statistics
}
\description{
Get Database Statistics
}
