% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-queries.R
\name{get_hansard_tables}
\alias{get_hansard_tables}
\title{Get Database Table References}
\usage{
get_hansard_tables(con)
}
\arguments{
\item{con}{Database connection}
}
\value{
List of table references
}
\description{
Returns dplyr table references for easy querying
}
