% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-queries.R
\name{get_top_speakers}
\alias{get_top_speakers}
\title{Get Top Speakers}
\usage{
get_top_speakers(con, limit = 10)
}
\arguments{
\item{con}{Database connection}

\item{limit}{Number of speakers to return}
}
\value{
Data frame with speaker statistics
}
\description{
Get Top Speakers
}
