% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansard_1850.R
\docType{data}
\name{hansard_1850}
\alias{hansard_1850}
\title{The 19th-century British Parliamentary debates for the decade 1850}
\format{
A data frame with 985755 rows and 2 variables:

sentence_id
A unique ID assigned to each sentence of the Hansard corpus.

text
A sentence spoken by a member of parliament.
}
\source{
\\href{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/ZCYJH8}{Harvard Dataverse}
}
\usage{
hansard_1850
}
\description{
Hansard corpus data for the decade 1850. To also access information about speaker, date, and debate title, join this data frame with debate_metadata on
sentence_id. The variables are as follows:
}
\examples{
data(hansard_1850)

}
\references{
Buongiorno, Steph; Kalescky, Robert; Godat, Eric; Cerpa, Omar Alexander; Guldi, Jo (2021)
(\\href{https://doi.org/10.7910/DVN/ZCYJH8})
}
\keyword{datasets}
