% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-data.R
\name{hansard_sample_info}
\alias{hansard_sample_info}
\title{List Available Sample Data}
\usage{
hansard_sample_info()
}
\value{
Data frame with information about available sample files
}
\description{
List Available Sample Data
}
\examples{
# See what sample data is available
hansard_sample_info()
}
