% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-import.R
\name{import_session_data}
\alias{import_session_data}
\title{Import Session Data}
\usage{
import_session_data(df, con, chamber, force_reimport = FALSE)
}
\arguments{
\item{df}{Data frame with session data}

\item{con}{Database connection}

\item{chamber}{Chamber type ("house" or "senate")}

\item{force_reimport}{Should existing session be replaced?}
}
\description{
Internal function to handle the actual database import
}
