% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-setup.R
\name{optimize_hansard_database}
\alias{optimize_hansard_database}
\title{Optimize Existing Database}
\usage{
optimize_hansard_database(con)
}
\arguments{
\item{con}{Database connection}
}
\description{
Apply optimizations to an existing database
}
