% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{parse_session_date}
\alias{parse_session_date}
\title{Parse Session Date from Filename}
\usage{
parse_session_date(filename)
}
\arguments{
\item{filename}{Filename to parse}
}
\value{
Date object or NA
}
\description{
Parse Session Date from Filename
}
