% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{parse_time}
\alias{parse_time}
\title{Parse Time Strings}
\usage{
parse_time(time_str)
}
\arguments{
\item{time_str}{Time string like "(12:04):"}
}
\value{
Cleaned time string or NA
}
\description{
Parse Time Strings
}
