% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-import.R
\name{process_debates}
\alias{process_debates}
\title{Process Debates Data}
\usage{
process_debates(df, con, session_id)
}
\arguments{
\item{df}{Session data frame}

\item{con}{Database connection}

\item{session_id}{Session ID}
}
\value{
Tibble mapping debateinfo to debate_id
}
\description{
Process Debates Data
}
