% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-import.R
\name{process_members}
\alias{process_members}
\title{Process Members Data}
\usage{
process_members(df, con, session_date)
}
\arguments{
\item{df}{Session data frame}

\item{con}{Database connection}

\item{session_date}{Date of session}
}
\value{
Tibble mapping name_id to member_id
}
\description{
Process Members Data
}
