% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-import.R
\name{process_speeches}
\alias{process_speeches}
\title{Process Speeches Data}
\usage{
process_speeches(df, con, session_id, member_map, debate_map)
}
\arguments{
\item{df}{Session data frame}

\item{con}{Database connection}

\item{session_id}{Session ID}

\item{member_map}{Member mapping}

\item{debate_map}{Debate mapping}
}
\description{
Process Speeches Data
}
