% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-search.R
\name{search_speech_content}
\alias{search_speech_content}
\title{Search Speech Content}
\usage{
search_speech_content(con, query, limit = 100, highlight = TRUE)
}
\arguments{
\item{con}{Database connection}

\item{query}{Search query string}

\item{limit}{Maximum number of results to return}

\item{highlight}{Whether to include highlighted snippets}
}
\value{
Data frame with search results
}
\description{
Full-text search of speech content using SQLite FTS5
}
