% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-validation.R
\name{validate_csv_batch}
\alias{validate_csv_batch}
\title{Validate Multiple Files}
\usage{
validate_csv_batch(file_paths)
}
\arguments{
\item{file_paths}{Vector of file paths}
}
\value{
Data frame with validation results
}
\description{
Validate Multiple Files
}
