% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-setup.R
\name{connect_hansard_database}
\alias{connect_hansard_database}
\title{Connect to Hansard Database}
\usage{
connect_hansard_database(db_path)
}
\arguments{
\item{db_path}{Path to the database file}
}
\value{
DBI connection object
}
\description{
Connect to Hansard Database
}
