% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-setup.R
\name{create_hansard_database}
\alias{create_hansard_database}
\title{Create Hansard Database}
\usage{
create_hansard_database(db_path, overwrite = FALSE)
}
\arguments{
\item{db_path}{Path to the database file}

\item{overwrite}{Should existing database be overwritten?}
}
\value{
DBI connection object
}
\description{
Creates a new SQLite database with the standard Hansard schema
}
