% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debate_metadata_1800.R
\docType{data}
\name{debate_metadata_1800}
\alias{debate_metadata_1800}
\title{The 19th-century British Parliamentary debates metadata for the decade 1800}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10 rows and 4 columns.
}
\source{
\\href{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/ZCYJH8}{Harvard Dataverse}
}
\usage{
debate_metadata_1800
}
\description{
The Hansard corpus debate metadata retains information about the debates such as the debate title and the speech date.
The variables are as follows:
}
\references{
Buongiorno, Steph; Kalescky, Robert; Godat, Eric; Cerpa, Omar Alexander; Guldi, Jo (2021)
(\\href{https://doi.org/10.7910/DVN/ZCYJH8})
}
\keyword{datasets}
