% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-processing.R
\name{import_hansard_batch}
\alias{import_hansard_batch}
\title{Import Multiple Files}
\usage{
import_hansard_batch(
  file_paths,
  con,
  chamber = c("house", "senate"),
  validate = TRUE,
  force_reimport = FALSE,
  progress = TRUE
)
}
\arguments{
\item{file_paths}{Vector of file paths}

\item{con}{Database connection}

\item{chamber}{Chamber type: "house" or "senate"}

\item{validate}{Should files be validated?}

\item{force_reimport}{Should existing sessions be overwritten?}

\item{progress}{Show progress bar?}
}
\value{
Summary of import results
}
\description{
Import Multiple Files
}
